edit doc, set default quantize_first_convolution to boolean
